+++ CHANGELOG.md
@@ -1,8 +1,26 @@
 # Changelog
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
-## 1.0.2
+## 1.2.0
 
+- minor: Install any subcharts
+- patch: update helm image version
+
+## 1.1.3
+
+- patch: Fix bug which causes error when trying to uninstall
+
+## 1.1.2
+
+- patch: Fix bug in uninstall schema
+
+## 1.1.1
+
+- patch: Update jinja version
+
+## 1.1.0
+
+- minor: Add uninstall flag
 - patch: Add support for --wait
 
 ## 1.0.1